requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dominate==2.7.0
Flask==2.2.3
Flask-Bootstrap==3.3.7.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
libusb1==3.0.0
MarkupSafe==2.1.2
ndeflib==0.3.3
nfcpy==1.0.4
pyDes==2.0.1
pyserial==3.5
python-dotenv==1.0.0
requests==2.28.2
urllib3==1.26.15
visitor==0.1.3
Werkzeug==2.2.3
